fix: separate the presence function in another file
